shiori/requirements.txt
discord.py==2.4.0
aiosqlite==0.17.0